Apply ../gramofile-1.6-mod/debian/patches/70-endian-fixes.dpatch from Debian's 1.6-9 release
DP: Fix endianness bugs in WAV and VOC headers on big-endian archs. DP: Use POSIX size types to avoid broken headers on 64bit archs. DP: Disable padding in structs that read/write raw on-disk data. DP: Merged with endianness patch by Tom Harvey <TomHarvey@IndustryFigure.com> [dk]

+++ endian.c
@@ -24,10 +24,10 @@ SwapTwo (short w)
   return (tmp);
 }
 
-u_long
-SwapFourBytes (u_long dw)
+u_int32_t
+SwapFourBytes (u_int32_t dw)
 {
-  register u_long tmp;
+  register u_int32_t tmp;
   tmp = (dw & 0x000000FF);
   tmp = ((dw & 0x0000FF00) >> 0x08) | (tmp << 0x08);
   tmp = ((dw & 0x00FF0000) >> 0x10) | (tmp << 0x08);
